move plugin to a hook, rename back to listener, fix notification logic and improve deployment detection
