+++ frontend/src/hooks/index.ts
@@ -8,3 +8,4 @@ export * from './useInputMap';
 export * from './useContractQueries';
 export * from './useFileName';
 export * from './useSetupStores';
+export * from './useTransactionListener';
