+++ frontend/src/main.ts
@@ -4,7 +4,7 @@ import { createPinia } from 'pinia';
 import Notifications from '@kyvg/vue3-notification';
 import App from './App.vue';
 import router from './router';
-import { persistStorePlugin, TransactionsListenerPlugin } from '@/plugins';
+import { persistStorePlugin } from '@/plugins';
 import { VueSpinnersPlugin } from 'vue3-spinners';
 import registerGlobalComponents from '@/components/global/registerGlobalComponents';
 import { VueQueryPlugin } from '@tanstack/vue-query';
@@ -32,7 +32,6 @@ app.use(FloatingVue, {
 });
 app.use(Notifications);
 app.use(VueSpinnersPlugin);
-app.use(TransactionsListenerPlugin);
 
 const plausible = createPlausible({
   init: {
