+++ frontend/src/plugins/index.ts
@@ -1,2 +1,1 @@
 export * from './persistStore';
-export * from './transactionsProcessor';
