+++ frontend/src/stores/contracts.ts
@@ -112,20 +112,12 @@ export const useContractsStore = defineStore('contractsStore', () => {
       (c) => c.contractId === contractId,
     );
 
-    const currentItem = deployedContracts.value[index];
     const newItem = { contractId, address, defaultState };
 
     if (index === -1) {
       deployedContracts.value.push(newItem);
     } else {
       deployedContracts.value.splice(index, 1, newItem);
-    }
-
-    if (currentItem.address !== newItem.address) {
-      notify({
-        title: 'Contract deployed',
-        type: 'success',
-      });
     }
   }
